feat: improve vue-query example to show maximum rendering time on SSR #85
