+++ examples/vue-query/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "strict": true,
-    "module": "ES2020",
+    "module": "ES2022",
     "moduleResolution": "Node",
     "target": "ES2020",
     "lib": ["DOM", "DOM.Iterable", "ESNext"],
